Add support for multiple Valets within the same access group (#297)
* Fix warning about missing global source in Gemfile

* Commit Xcode-generated changes

These changes happened automatically as a result of using a newer Xcode version

* Add support for multiple Valets within the same access group

Previously, the shared access group identifier was used for both the access group itself, as well as the uniqueness identifier for the given Valet. This adds an optional additional `identifier` parameter that can be specified when creating a shared access group Valet. The identifier adds an additional element of uniqueness, so two Valets with the same shared access group can exist, and their data will not overlap, so long as they have different identifiers.

The default for this value is `nil`, which keeps the existing behavior for full backward compatibility.

* Bump version to 4.2.0

---------

Co-authored-by: Dan Federman <dfed@me.com>